Rename proximal module to proximal_lang.py
